chore: class renaming
