+++ src/electron/platform/android/atfa-scan-results-to-unified-rules.ts
