+++ src/electron/platform/android/axe-scan-results-to-unified-results.ts
