+++ src/electron/platform/android/scan-results-to-unified-results.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 import { UnifiedResult } from 'common/types/store-data/unified-data-interface';
 import { UUIDGenerator } from 'common/uid-generator';
-import { convertAtfaScanResultsToUnifiedResults } from 'electron/platform/android/atfa-results-to-unified-results';
-import { convertAxeScanResultsToUnifiedResults } from 'electron/platform/android/axe-results-to-unified-results';
+import { convertAtfaScanResultsToUnifiedResults } from 'electron/platform/android/atfa-scan-results-to-unified-results';
+import { convertAxeScanResultsToUnifiedResults } from 'electron/platform/android/axe-scan-results-to-unified-results';
 import { AndroidScanResults } from './android-scan-results';
 import { RuleInformationProviderType } from './rule-information-provider-type';
